Move legendre2-code to legendre
Remove old legendre code

Use internal comparison function in zeeman/ycalc.py

Make zeeman/ycalc.py fast
