+++ controlfiles-python/CMakeLists.txt
@@ -33,4 +33,4 @@ if (ARTS_XML_DATA_DIR)
   arts_test_run_pyfile(slow artscomponents/surfaceblackbody/TestSurfaceBlackbody.py)
 endif()
 
-arts_test_run_pyfile(slow artscomponents/zeeman/ycalc.py)
+arts_test_run_pyfile(fast artscomponents/zeeman/ycalc.py)
